[swappable] Improvements to #50
