+++ openwisp_monitoring/device/admin.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from django.utils.html import format_html
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
-from swapper import is_swapped, load_model, split
+from swapper import load_model
 
 from openwisp_controller.config.admin import DeviceAdmin as BaseDeviceAdmin
 from openwisp_controller.config.models import Device
@@ -21,11 +21,7 @@ class DeviceAdmin(BaseDeviceAdmin):
         ctx = super().get_extra_context(pk)
         if pk:
             device_data = DeviceData(pk=pk)
-            if is_swapped('monitoring', 'Metric'):
-                monitoring_app = split(is_swapped('monitoring', 'Metric'))[0]
-            else:
-                monitoring_app = 'monitoring'
-            api_url = reverse(f'{monitoring_app}:api_device_metric', args=[pk])
+            api_url = reverse(f'monitoring:api_device_metric', args=[pk])
             ctx.update(
                 {
                     'device_data': device_data.data_user_friendly,
