+++ openwisp_monitoring/device/api/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url
 
 from . import views
 
-app_name = 'openwisp_monitoring.device'
+app_name = 'monitoring'
 
 urlpatterns = [
     url(
