+++ openwisp_monitoring/monitoring/tests/test_graphs.py
@@ -1,12 +1,11 @@
 import json
 from datetime import date, timedelta
-from unittest import skipIf
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 from django.utils.timezone import now
-from swapper import is_swapped, load_model
+from swapper import load_model
 
 from . import TestMonitoringMixin
 
@@ -213,7 +212,6 @@ class TestGraphs(TestMonitoringMixin, TestCase):
         g = self._create_graph(test_data=False)
         self.assertEqual(g.description, 'Dummy chart for testing purposes.')
 
-    @skipIf(is_swapped('monitoring', 'Metric'), 'Running tests on sample_app')
     def test_wifi_hostapd(self):
         m = self._create_object_metric(
             name='wifi associations', key='hostapd', field_name='mac_address'
