+++ tests/openwisp2/sample_device_monitoring/tests.py
@@ -1,7 +1,9 @@
 from django.contrib.auth import get_user_model
 from django.test import TestCase
-from openwisp_monitoring.device.tests import DeviceMonitoringTestCase
-from openwisp_monitoring.device.tests.test_models import BaseTestCase
+from openwisp_monitoring.device.tests import (
+    DeviceMonitoringTestCase as DeviceMonitoringTestCase,
+)
+from openwisp_monitoring.device.tests.test_api import TestDeviceApi as BaseTestDeviceApi
 from openwisp_monitoring.device.tests.test_models import (
     TestDeviceData as BaseTestDeviceData,
 )
@@ -16,19 +18,19 @@ from openwisp_monitoring.device.tests.test_settings import (
 )
 
 
-class TestRecovery(BaseTestRecovery, DeviceMonitoringTestCase):
+class TestRecovery(BaseTestRecovery):
     pass
 
 
-class TestDeviceData(BaseTestDeviceData, BaseTestCase):
+class TestDeviceData(BaseTestDeviceData):
     pass
 
 
-class TestDeviceMonitoring(BaseTestDeviceMonitoring, BaseTestCase):
+class TestDeviceMonitoring(BaseTestDeviceMonitoring):
     pass
 
 
-class TestSettings(BaseTestSettings, DeviceMonitoringTestCase):
+class TestSettings(BaseTestSettings):
     pass
 
 
@@ -43,3 +45,16 @@ class TestAdmin(TestCase):
         self._login_admin()
         r = self.client.get('/admin/')
         self.assertContains(r, '/admin/sample_device_monitoring/detailsmodel/')
+
+
+class TestDeviceApi(BaseTestDeviceApi):
+    pass
+
+
+del BaseTestRecovery
+del DeviceMonitoringTestCase
+del BaseTestDeviceData
+del BaseTestDeviceMonitoring
+del TestAdmin
+del BaseTestSettings
+del BaseTestDeviceApi
